sources/backend/pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["yhoshi <francekakumei714@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.7"
djangorestframework = "^3.14.0"
drf-spectacular = "^0.26.5"
mysqlclient = "^2.2.0"
django-boost = "^2.1"
black = "^23.11.0"
ruff = "^0.1.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
